add enableBaubles option to Backpacks
