+++ src/main/java/vazkii/quark/oddities/feature/Backpacks.java
@@ -44,17 +44,19 @@ import net.minecraftforge.items.ItemHandlerHelper;
 import vazkii.arl.network.NetworkHandler;
 import vazkii.arl.recipe.RecipeHandler;
 import vazkii.arl.util.InventoryIIH;
+import vazkii.quark.base.module.ConfigHelper;
 import vazkii.quark.base.module.Feature;
 import vazkii.quark.base.network.message.MessageHandleBackpack;
 import vazkii.quark.oddities.RecipesBackpackDyes;
 import vazkii.quark.oddities.client.gui.GuiBackpackInventory;
 import vazkii.quark.oddities.item.ItemBackpack;
+import vazkii.quark.oddities.item.ItemBackpackBaubles;
 
 public class Backpacks extends Feature {
 
 	public static ItemBackpack backpack;
 	
-	public static boolean superOpMode, enableTrades, enableCrafting, enablePickUp, ignoreShiftClick;
+	public static boolean superOpMode, enableTrades, enableCrafting, enablePickUp, ignoreShiftClick, enableBaubles;
 
 	public static  int leatherCount, minEmeralds, maxEmeralds;
 	
@@ -74,11 +76,12 @@ public class Backpacks extends Feature {
 		leatherCount = loadPropInt("Required Leather", "", 12);
 		minEmeralds = loadPropInt("Min Required Emeralds", "", 12);
 		maxEmeralds = loadPropInt("Max Required Emeralds", "", 18);
+		if (Loader.isModLoaded("baubles")) enableBaubles = loadPropBool("Enable Backpack Bauble", "Set this to true to turn the backpack into a bauble that's equipped in the Body slot", true);
 	}
 	
 	@Override
 	public void preInit(FMLPreInitializationEvent event) {
-		backpack = new ItemBackpack();
+		backpack = ((Loader.isModLoaded("baubles") && enableBaubles) ? new ItemBackpackBaubles() : new ItemBackpack());
 		
 		if (enableCrafting)
 			RecipeHandler.addOreDictRecipe(new ItemStack(backpack),
