+++ src/main/java/vazkii/quark/oddities/item/ItemBackpack.java
@@ -61,22 +61,20 @@ import net.minecraftforge.items.ItemStackHandler;
 import vazkii.arl.interf.IItemColorProvider;
 import vazkii.arl.item.ItemModArmor;
 import vazkii.arl.util.ItemNBTHelper;
+import vazkii.quark.base.Quark;
 import vazkii.quark.base.handler.ProxiedItemStackHandler;
 import vazkii.quark.base.item.IQuarkItem;
+import vazkii.quark.base.lib.LibGuiIDs;
 import vazkii.quark.base.lib.LibMisc;
 import vazkii.quark.oddities.client.model.ModelBackpack;
 import vazkii.quark.oddities.feature.Backpacks;
 
-@Optional.InterfaceList(value = {
-	@Optional.Interface(iface = "baubles.api.IBauble", modid = "baubles", striprefs = true),
-	@Optional.Interface(iface = "baubles.api.render.IRenderBauble", modid = "baubles", striprefs = true)
-})
-public class ItemBackpack extends ItemModArmor implements IBauble, IQuarkItem, IItemColorProvider, IRenderBauble {
+public class ItemBackpack extends ItemModArmor implements IQuarkItem, IItemColorProvider {
 	private static final String WORN_TEXTURE = LibMisc.PREFIX_MOD + "textures/misc/backpack_worn.png";
 	private static final String WORN_OVERLAY_TEXTURE = LibMisc.PREFIX_MOD + "textures/misc/backpack_worn_overlay.png";
 	
-	private static final ResourceLocation WORN_TEXTURE_RL = new ResourceLocation(WORN_TEXTURE);
-	private static final ResourceLocation WORN_OVERLAY_TEXTURE_RL = new ResourceLocation(WORN_OVERLAY_TEXTURE);
+	protected static final ResourceLocation WORN_TEXTURE_RL = new ResourceLocation(WORN_TEXTURE);
+	protected static final ResourceLocation WORN_OVERLAY_TEXTURE_RL = new ResourceLocation(WORN_OVERLAY_TEXTURE);
 
 	public static final IBehaviorDispenseItem DISPENSER_BEHAVIOR = new BehaviorDefaultDispenseItem() {
 		/**
@@ -162,32 +160,15 @@ public class ItemBackpack extends ItemModArmor implements IBauble, IQuarkItem, I
 	public ActionResult<ItemStack> onItemRightClick(World world, EntityPlayer player, EnumHand hand) {
 		ItemStack stack = player.getHeldItem(hand);
 		if (!world.isRemote) {
-			if (Loader.isModLoaded("baubles")) {
-				IBaublesItemHandler baubles = BaublesApi.getBaublesHandler(player);
-				for (int i = 0; i < baubles.getSlots(); i++) {
-					if ((baubles.getStackInSlot(i) == null || baubles.getStackInSlot(i).isEmpty())
-							&& baubles.isItemValidForSlot(i, stack, player)) {
-						baubles.setStackInSlot(i, stack.copy());
-						if (!player.capabilities.isCreativeMode) {
-							player.inventory.setInventorySlotContents(player.inventory.currentItem, ItemStack.EMPTY);
-						}
-						onEquipped(stack, player);
-						
-						return new ActionResult<ItemStack>(EnumActionResult.SUCCESS, stack);
-					}
-				}
-				return new ActionResult<ItemStack>(EnumActionResult.FAIL, stack);
-			} else {
-				EntityEquipmentSlot entityequipmentslot = EntityLiving.getSlotForItemStack(stack);
-				ItemStack stackInSlot = player.getItemStackFromSlot(entityequipmentslot);
+			EntityEquipmentSlot entityequipmentslot = EntityLiving.getSlotForItemStack(stack);
+			ItemStack stackInSlot = player.getItemStackFromSlot(entityequipmentslot);
 
-				if (stackInSlot.isEmpty()) {
-					player.setItemStackToSlot(entityequipmentslot, stack.copy());
-					stack.setCount(stack.getCount()-1);
-					return new ActionResult<ItemStack>(EnumActionResult.SUCCESS, stack);
-				} else {
-					return new ActionResult<ItemStack>(EnumActionResult.FAIL, stack);
-				}
+			if (stackInSlot.isEmpty()) {
+				player.setItemStackToSlot(entityequipmentslot, stack.copy());
+				stack.setCount(stack.getCount()-1);
+				return new ActionResult<ItemStack>(EnumActionResult.SUCCESS, stack);
+			} else {
+				return new ActionResult<ItemStack>(EnumActionResult.FAIL, stack);
 			}
 		}
 		return new ActionResult<ItemStack>(EnumActionResult.SUCCESS, stack);
@@ -305,24 +286,6 @@ public class ItemBackpack extends ItemModArmor implements IBauble, IQuarkItem, I
 		return (stack, i) -> i == 1 ? ((ItemBackpack) stack.getItem()).getColor(stack) : -1;
 	}
 	
-	@Override
-	@Optional.Method(modid = "baubles")
-	public BaubleType getBaubleType(ItemStack itemStack) {
-		return BaubleType.BODY;
-	}
-	
-	@Override
-	@Optional.Method(modid = "baubles")
-	public boolean canEquip(ItemStack itemstack, EntityLivingBase player) {
-		return !Backpacks.isEntityWearingBackpack(player, itemstack);
-	}
-	
-	@Override
-	@Optional.Method(modid = "baubles")
-	public boolean canUnequip(ItemStack itemstack, EntityLivingBase player) {
-		return Backpacks.superOpMode || !doesBackpackHaveItems(itemstack);
-	}
-	
 	public boolean hasColor(@Nonnull ItemStack stack) {
 		NBTTagCompound nbttagcompound = stack.getTagCompound();
 		return (nbttagcompound != null)
@@ -359,51 +322,6 @@ public class ItemBackpack extends ItemModArmor implements IBauble, IQuarkItem, I
 		}
 		
 		nbttagcompound1.setInteger("color", color);
-	}
-	
-	@Override
-	@SideOnly(Side.CLIENT)
-	@Optional.Method(modid = "baubles")
-	public void onPlayerBaubleRender(ItemStack itemStack, EntityPlayer player, RenderType renderType, float v) {
-		if (!player.world.isRemote) return;
-
-		if (renderType != RenderType.BODY) return;
-
-		if (model == null) model = new ModelBackpack();
-		
-		model.setModelAttributes(new ModelPlayer(0.0F, false));
-		
-		Minecraft.getMinecraft().renderEngine.bindTexture(WORN_TEXTURE_RL);
-		
-		int i = backpack.getColor(itemStack);
-		float red = (float) (i >> 16 & 255) / 255.0F;
-		float green = (float) (i >> 8 & 255) / 255.0F;
-		float blue = (float) (i & 255) / 255.0F;
-		
-		GlStateManager.color(red, green, blue, 1);
-		
-
-		float partialTicks = Minecraft.getMinecraft().getRenderPartialTicks();
-		float f = this.interpolateRotation(player.prevRenderYawOffset, player.renderYawOffset, partialTicks);
-		float f1 = this.interpolateRotation(player.prevRotationYawHead, player.rotationYawHead, partialTicks);
-		float f2 = f1 - f;
-//		float f4 = renderer.prepareScale(player, Minecraft.getMinecraft().getRenderPartialTicks());
-		float f4 = 0.0625F;
-		float f8 = (float) player.ticksExisted + partialTicks;
-		float f5 = player.prevLimbSwingAmount + (player.limbSwingAmount - player.prevLimbSwingAmount) * partialTicks;
-		float f6 = player.limbSwing - player.limbSwingAmount * (1.0F - partialTicks);
-		float f7 = player.prevRotationPitch + (player.rotationPitch - player.prevRotationPitch) * partialTicks;
-		
-		model.setVisible(false);
-		model.bipedBody.showModel = true;
-		
-		model.render(player, f6, f5, f8, f2, f7, f4);
-		
-		Minecraft.getMinecraft().renderEngine.bindTexture(WORN_OVERLAY_TEXTURE_RL);
-		
-		GlStateManager.color(1, 1, 1, 1);
-		
-		model.render(player, 0, 0, 1000, 0, 0, 0.0625F);
 	}
 	
 	/**
